refactor: correct precision logic and refine dynamic string APIs
- Correct logical errors in precision calculation for printf and strfrom backends.
- Use raw precision sentinel to properly distinguish between unset and explicit zero precision.
- Fix signedness bug in custom formatter rounding logic by casting before negating precision.
- Remove redundant casts and update precision variable types to size_t.
- Finalize naming for dynamic string macros: yio_asprintf, yio_asprint, yio_append, yio_append_s.
- Simplify pfmt_parse test with static zero-initialized structure.

+++ src/yio/private/yio_float_strfrom_ryu.c
@@ -145,7 +145,7 @@ int YYIO_float_astrfrom_ryu$1_in(YYIO_string *res, int precision0, char spec, YY
 	}
 	// https://github.com/ulfjack/ryu/issues/197
 	const size_t minsize = 2000;
-	const int precision = yio_precision_get_default(precision0, 6);
+	const size_t precision = yio_precision_get_default(precision0, 6);
 	const size_t toalloc = minsize + precision + 1;
 	int err = YYIO_string_reserve(res, toalloc);
 	if (err) return err;
